requirements.txt
Django==5.1.6
djangorestframework==3.15.2
requests==2.32.3
python-dotenv==1.0.1
dj-database-url==2.3.0
